Fix click event bleed-through when clicking on closed widget
Select2 is supposed to trap all mouse events inside the select2-drop element,
but this was not 100% working on FireFox.
After much debugging and head-scratching, it turned out that a legacy IE
workaround was masking the problem; removing the workaround caused the bug
to consistently show on all browsers, and the solution was to target click
events direcly in order to cancel them.

See https://lab.civicrm.org/dev/core/-/issues/5932